requirements.txt
langgraph
tavily-python
langchain
langchain-anthropic
pypandoc
pandoc
tldextract
fpdf
flask
fastapi 
uvicorn
python-multipart
python-dotenv
markdown2


